+++ tests/Rest/GroupProviderTest.php
@@ -66,10 +66,26 @@ class GroupProviderTest extends AbstractGroupControllerTestCase
 
     public function testGetGroupCollection(): void
     {
+        // current user has manage grants for groups 1 and 2, a read grant for group 3, and a write grant for group 5
+        $this->login(self::CURRENT_USER_IDENTIFIER);
         $group1 = $this->addTestGroupAndManageGroupGrantForCurrentUser(self::TEST_GROUP_NAME);
         $group2 = $this->addTestGroupAndManageGroupGrantForCurrentUser(self::TEST_GROUP_NAME);
-        $group3 = $this->addTestGroupAndManageGroupGrantForCurrentUser(self::TEST_GROUP_NAME);
 
+        // another user has manage grants for groups 3, 4, and 5
+        $this->login(self::ANOTHER_USER_IDENTIFIER);
+        $group3 = $this->testEntityManager->addGroup(self::TEST_GROUP_NAME);
+        $group3ManageGrant = $this->authorizationService->addGroup($group3->getIdentifier());
+        $this->testEntityManager->addResourceActionGrant($group3ManageGrant->getAuthorizationResource(),
+            AuthorizationService::READ_GROUP_ACTION, self::CURRENT_USER_IDENTIFIER);
+
+        $group4 = $this->addTestGroupAndManageGroupGrantForCurrentUser(self::TEST_GROUP_NAME);
+
+        $group5 = $this->testEntityManager->addGroup(self::TEST_GROUP_NAME);
+        $group5ManageGrant = $this->authorizationService->addGroup($group5->getIdentifier());
+        $this->testEntityManager->addResourceActionGrant($group5ManageGrant->getAuthorizationResource(),
+            AuthorizationService::DELETE_GROUP_ACTION, self::CURRENT_USER_IDENTIFIER);
+
+        $this->login(self::CURRENT_USER_IDENTIFIER);
         $groups = $this->groupProviderTester->getCollection([
             'page' => 1,
             'perPage' => 10,
@@ -94,5 +110,15 @@ class GroupProviderTest extends AbstractGroupControllerTestCase
         ]);
         $this->assertCount(1, $groups);
         $this->assertEquals($group3->getIdentifier(), $groups[0]->getIdentifier());
+
+        $this->login(self::ANOTHER_USER_IDENTIFIER);
+        $groups = $this->groupProviderTester->getCollection([
+            'page' => 1,
+            'perPage' => 10,
+        ]);
+        $this->assertCount(3, $groups);
+        $this->assertEquals($group3->getIdentifier(), $groups[0]->getIdentifier());
+        $this->assertEquals($group4->getIdentifier(), $groups[1]->getIdentifier());
+        $this->assertEquals($group5->getIdentifier(), $groups[2]->getIdentifier());
     }
 }